refactor(mapgen-studio): extract config overrides feature
Move config overrides UI/state into features/configOverrides.
Wire App.tsx to useConfigOverrides + ConfigOverridesPanel.
Preserve schema normalization + JSON validation behavior.
