+++ apps/mapgen-studio/src/features/configOverrides/useConfigOverrides.ts
@@ -1,0 +1,102 @@
+import { useCallback, useMemo, useState } from "react";
+import { validateConfigOverridesJson } from "./validate";
+
+export type ConfigOverridesTab = "form" | "json";
+
+export type UseConfigOverridesArgs<TConfig> = {
+  baseConfig: TConfig;
+  schema: unknown;
+  disabled?: boolean;
+  basePathForErrors?: string;
+};
+
+export type UseConfigOverridesResult<TConfig> = {
+  enabled: boolean;
+  setEnabled(next: boolean): void;
+
+  tab: ConfigOverridesTab;
+  setTab(next: ConfigOverridesTab): void;
+
+  value: TConfig;
+  setValue(next: TConfig): void;
+
+  jsonText: string;
+  setJsonText(next: string): void;
+  jsonError: string | null;
+
+  reset(): void;
+  applyJson(): { ok: boolean };
+
+  configOverridesForRun: TConfig | undefined;
+};
+
+export function useConfigOverrides<TConfig>(
+  args: UseConfigOverridesArgs<TConfig>
+): UseConfigOverridesResult<TConfig> {
+  const { baseConfig, schema, basePathForErrors = "/configOverrides" } = args;
+  const [enabled, setEnabled] = useState(false);
+  const [tab, setTabState] = useState<ConfigOverridesTab>("form");
+  const [value, setValueState] = useState<TConfig>(baseConfig);
+  const [jsonText, setJsonText] = useState(() => JSON.stringify(baseConfig, null, 2));
+  const [jsonError, setJsonError] = useState<string | null>(null);
+
+  const reset = useCallback(() => {
+    setValueState(baseConfig);
+    setJsonText(JSON.stringify(baseConfig, null, 2));
+    setJsonError(null);
+  }, [baseConfig]);
+
+  const applyJson = useCallback((): { ok: boolean } => {
+    const result = validateConfigOverridesJson<TConfig>(schema, jsonText, basePathForErrors);
+    if (!result.ok) {
+      setJsonError(result.error);
+      return { ok: false };
+    }
+    setValueState(result.value);
+    setJsonError(null);
+    return { ok: true };
+  }, [schema, jsonText, basePathForErrors]);
+
+  const setValue = useCallback((next: TConfig) => {
+    setValueState(next);
+    setJsonError(null);
+  }, []);
+
+  const setTab = useCallback(
+    (next: ConfigOverridesTab) => {
+      if (next === tab) return;
+      if (next === "form") {
+        const result = validateConfigOverridesJson<TConfig>(schema, jsonText, basePathForErrors);
+        if (!result.ok) {
+          setJsonError(result.error);
+          return;
+        }
+        setValueState(result.value);
+        setJsonError(null);
+        setTabState("form");
+        return;
+      }
+      setJsonText(JSON.stringify(value, null, 2));
+      setJsonError(null);
+      setTabState("json");
+    },
+    [tab, schema, jsonText, basePathForErrors, value]
+  );
+
+  const configOverridesForRun = useMemo(() => (enabled ? value : undefined), [enabled, value]);
+
+  return {
+    enabled,
+    setEnabled,
+    tab,
+    setTab,
+    value,
+    setValue,
+    jsonText,
+    setJsonText,
+    jsonError,
+    reset,
+    applyJson,
+    configOverridesForRun,
+  };
+}
